okx-bot_reserve/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=0.20.0
pyyaml>=6.0
matplotlib>=3.6.0
scipy>=1.9.0
